rematar proyecto

+++ dwes/header.php
@@ -1,6 +1,6 @@
 <?php
 session_start();
-
+// print_r($_SESSION['rols']);
 
 ?>
 
@@ -28,23 +28,33 @@ session_start();
       <nav>
         <ul class="flex space-x-4">
           <li><a href="/student067/dwes/index.php" class="text-white hover:text-gray-300">Inicio</a></li>
+          <?php if(isset($_SESSION['rols'])) { ?>
           <li><p class="text-white hover:text-gray-300">Customers</p><ul>
+          <?php  if(!in_array("customer",$_SESSION['rols'][0]) || (in_array("admin",$_SESSION['rols'][0]) || in_array("employee",$_SESSION['rols'][0]) )) {  ?>
             <li><a href="/student067/dwes/pages/forms/form_customer_select.php" class="text-white hover:text-gray-300">Select Customer</a></li>
-            <li><a href="/student067/dwes/pages/forms/form_customer_insert.php" class="text-white hover:text-gray-300">Insert Customer</a></li>
+       
+              <li><a href="/student067/dwes/pages/forms/form_customer_insert.php" class="text-white hover:text-gray-300">Insert Customer</a></li>
+           <?php } ?>
             <li><a href="/student067/dwes/pages/forms/form_customer_delete.php" class="text-white hover:text-gray-300">Delete Customer</a></li>
             <li><a href="/student067/dwes/pages/forms/form_customer_update_call_id.php"class="text-white hover:text-gray-300" >Update Customer</a></li>
           </ul></li>
+          <?php }  ?>
+          <?php if(isset($_SESSION['rols'])) { ?>
+          <?php  if(!in_array("customer",$_SESSION['rols'][0]) || (in_array("admin",$_SESSION['rols'][0]) || in_array("employee",$_SESSION['rols'][0]) )) {  ?>
           <li><a href="#" class="text-white hover:text-gray-300">Rooms</a><ul>
             <li><a href="/student067/dwes/pages/forms/form_place_select.php" class="text-white hover:text-gray-300">Select Room</a></li>
             <li><a href="/student067/dwes/pages/forms/form_place_insert.php" class="text-white hover:text-gray-300">Insert Room</a></li>
             <li><a href="/student067/dwes/pages/forms/form_place_delete.php" class="text-white hover:text-gray-300">Delete Room</a></li>
             <li><a href="/student067/dwes/pages/forms/form_place_update_call_id.php" class="text-white hover:text-gray-300">Update Room</a></li>
           </ul></li>
-          <li><a href="#" class="text-white hover:text-gray-300">Reservations</a><ul>
-            <li><a href="/student067/dwes/pages/forms/form_reservation_select.php" class="text-white hover:text-gray-300">Show Reservations</a></li>
+          <?php }} ?> 
+          <?php if(isset($_SESSION['rols'])) { ?>
+          <li><a href="#" class="text-white hover:text-gray-300">Reservations</a><ul>   
+            <li><a href="/student067/dwes/pages/forms/form_reservation_select.php" class="text-white hover:text-gray-300">Show Reservations</a></li> 
             <li><a href="/student067/dwes/pages/forms/form_reservation_choose_date.php" class="text-white hover:text-gray-300">Do Reservations</a></li>
             <li><a href="/student067/dwes/pages/forms/form_reservation_delete.php" class="text-white hover:text-gray-300">Delete Reservations</a></li>
             <li><a href="/student067/dwes/pages/forms/form_reservation_call_id.php" class="text-white hover:text-gray-300">Update Reservations</a></li>
+            <?php } ?>
           </ul></li>
         </ul>
       </nav>
